requirements.txt
appdirs==1.4.4
black==20.8b1
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
keyboard==0.13.5
MouseInfo==0.1.3
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.1.2
pkg-resources==0.0.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
python-libxdo==0.1.2a1
python3-xlib==0.15
PyTweening==1.0.3
regex==2020.11.13
requests==2.25.1
six==1.15.0
standardjson==0.3.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
webhooks==0.4.2
websockets==8.1
wrapt==1.12.1
